techdocs-node: inline azure mocks
